fix(server): pair memConfig entries with the same greedy claimer
The mirror looked its config entry up with a first-match sameIdentity
scan, the pattern the surrounding function was written to avoid.
sameIdentity compares org keys only when both sides carry one, and a
memConfig entry never receives the org backfill, so two entries sharing
an accountUuid both matched the first one; two apikey entries sharing a
name collide the same way through the name fallback. The binding then
landed on the wrong account, and the next save persisted it to disk.

Claim config entries greedily in disk order, mirroring the manager-side
bookkeeping. Identity pairing also survives the two lists holding
different numbers of entries, which they do whenever a credential-less
account is dropped at startup.

Move the function to its own module so the pairing is reachable from a
test without executing the CLI.
